Added gen_database.py
Decided that this project will cosist of (for now) two main files: one
for generating the db and one for querying and getting information
out of the generated db.

gen_database.py contains the functions to build the database and its
tables. Currently it generates the file table which contains info on all
the files.

filetest.py will continue to be my messing around file to get the
functions ready for gen_database.py.