Scraping works / Added queries file
Scraping of the web source files works now and builds both the link and
broken [link] tables. Need to push a couple things off to
a configuration file, but things all work for now.

Also added a queries file that will contain useful queries as I come up
with them.