More better link table generation
Added special case function for my dataset: I'm treating index files as
root nodes. However, most links are actually in a menu file that is
included via php function rather than a href html link. This means that
the link entries for index files will have very few items because the
menu is really where all the links are.

So, to solve that, I made a special case function that inserts a link
entry pointing index -> menu. This will allow the dfs algo to work
properly because now links are visible in the tree.

Added link checking to look for actions on form tags.

Due to above, needed to pop out code for link regex checking so that
Paxton wouldn't be mad that I reused code.

Added the index -> menu query to the queries file.